can-roundtrip-stats: a tool to benchmark transmission time
This is a simple tool I wrote in the past. It will report the time
need for a packet to travel from:
  * application TX path to kernel TX path
  * kernel TX path to kernel RX path (i.e. kernel round trip)
  * kernel RX path to application RX path
  * application TX path to application RX path (i.e application round
    trip)

This tool is useful to benchmark latency on software and hardware. It
can be used to, for example:
  * compare performances of two CAN controllers (by using the kernel
    round trip information)
  * compare the different CAN queue disciplines (by using the
    application TX path to kernel TX path information)

I am sharing it as-is. Please see this message as an FYI. I do not
consider this mature enough and I am not expecting anyone to pick that
patch. Everything is hard coded, I did not put effort to make it
configurable.

The tool requires the TX timestamps (which I previously added to the
kernel in [1] and [2]).

My ultimate goal was to add the TX timestamp feature to candump from
can-utils [3], however, I faced a blocker: the TX timestamps are sent
to the error queue of the sender's socket meaning that a listener
(such as candump) will not see those TX timestamp on his error queue
because this is a different socket as the sender. If anyone knows a
method to access the error queue of another process's socket, let me
know (I guess that you need to be root, but did not find how to do
it).

Because I did not manage to add the feature to candump, I am sharing
instead this standalone tool, hoping someone might find it useful.

At the moment, I am not planning to invest more time in the
foreseeable future. If someone want to take over and make is a bit
more sexy so that it can reach can-utils, go ahead. I think that it
basically misses a command line interface in the same fashion of
cangen to make is configurable.

Else, you can also tweak the hard coded information to easily adjust
it to your needs.

[1] commit eb88531bdbfa ("can: raw: add missing error queue support")
Link: https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=eb88531bdbfaafb827192d1fc6c5a3fcc4fadd96

[2] commit 741b91f1b0ea ("can: dev: can_put_echo_skb(): add software tx timestamps")
Link: https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=741b91f1b0ea34f00f6a7d4539b767c409291fcf

[3] candump.c
Link: https://github.com/linux-can/can-utils/blob/master/candump.c

Link: https://lore.kernel.org/all/20220626075317.746535-1-mailhol.vincent@wanadoo.fr
Signed-off-by: Vincent Mailhol <mailhol.vincent@wanadoo.fr>
Signed-off-by: Marc Kleine-Budde <mkl@pengutronix.de>